move last day logic out to calendar index
