+++ src/features/calendar/components/CalendarDayView/index.tsx
@@ -1,31 +1,21 @@
 import Box from '@mui/material/Box';
-import { CampaignActivity } from 'features/campaigns/models/CampaignActivitiesModel';
-import dayjs from 'dayjs';
 
 import Day from './Day';
+import { DaySummary } from '../utils';
 import PreviousDayPrompt from './PreviousDayPrompt';
-import { getActivitiesByDate, getPreviousDay } from './utils';
 
 export interface CalendarDayViewProps {
-  activities: CampaignActivity[];
+  activities: [string, DaySummary][];
+  previousActivityDay: [Date, DaySummary] | null;
   focusDate: Date;
   onClickPreviousDay: (date: Date) => void;
 }
 
 const CalendarDayView = ({
   activities,
-  focusDate,
+  previousActivityDay,
   onClickPreviousDay,
 }: CalendarDayViewProps) => {
-  const activitiesByDate = getActivitiesByDate(activities);
-  const futureDays = Object.entries(activitiesByDate).filter(([dateString]) => {
-    const activityDay = dayjs(dateString);
-    return (
-      activityDay.isSame(focusDate, 'day') || activityDay.isAfter(focusDate)
-    );
-  });
-  const previousActivityDay = getPreviousDay(activitiesByDate, focusDate);
-
   return (
     <Box display="flex" flexDirection="column" gap={2}>
       {previousActivityDay && (
@@ -38,7 +28,7 @@ const CalendarDayView = ({
         />
       )}
       {/* List of days with events */}
-      {futureDays.map(([dateString, daySummary], index) => {
+      {activities.map(([dateString, daySummary], index) => {
         return (
           <Day key={index} date={new Date(dateString)} dayInfo={daySummary} />
         );
