--- src/features/calendar/components/types.ts
@@ -1,5 +1,0 @@
-export enum TimeScale {
-  DAY = 'day',
-  WEEK = 'week',
-  MONTH = 'month',
-}
